[ENH] traction: default to OdooBot for Issue and Headline assignee
When the user selects 'Issue' or 'Headline' as the type for a new
activity, the assigned user changes to OdooBot (root user). This is done
to avoid having activities pile up in users' todo list when they're
really just headlines to discuss or issues to solve at a given team
meeting.