[DEV] Massive changes to the structure of issues and events.
Added a button to the systray to add issues from anywhere. Need to clean
it up as the standard form view sucks. Vision is to turn this into a
simple set of OWL components instead of using built-in form view pop-up
for creation of issues. Probably re-use it for headlines as well.

Having trouble grabbing the current record so may decide to leave this
functionality of linking to a specific record for now.